Tried to improve a user's first contact with the site:
Tooltips are now manipulated programmatically, so that there will not be unecessary flashing when you move the mouse.  It also won't flash anymove when the text updates.  Autocomplete and some other options are turned off.  You can submit your username with clicking, or pressing enter.  The tooltip has been styled to be red.